+++ lib/src/cassandart_impl.dart
@@ -8,14 +8,17 @@ import 'package:buffer/buffer.dart';
 import 'package:meta/meta.dart';
 import 'package:page/page.dart';
 
-part 'cassandra_pool.dart';
+part 'cluster.dart';
 part 'collection.dart';
 part 'frame_protocol.dart';
 part 'frames.dart';
 part 'queries.dart';
 part 'serialization.dart';
 
-abstract class CassandraClient {
+/// The execution and query context of the Cassandra client. Underlying
+/// implementation may use a connection pool or just a single connection.
+abstract class Client {
+  /// Execute [query] with the given parameters.
   Future execute(
     String query, {
     Consistency consistency,
@@ -23,6 +26,8 @@ abstract class CassandraClient {
     values,
   });
 
+  /// Execute data row [query] with the given parameters and return a page
+  /// object of the results rows (and further pagination support).
   Future<RowsPage> query(
     String query, {
     Consistency consistency,
@@ -33,6 +38,7 @@ abstract class CassandraClient {
   });
 }
 
+/// The consistency of an operation.
 enum Consistency {
   any,
   one,
@@ -47,10 +53,13 @@ enum Consistency {
   localOne,
 }
 
+/// Provides response to authentication challenges.
 abstract class Authenticator {
+  /// Responds to the authentication [challenge].
   Future<Uint8List> respond(Uint8List challenge);
 }
 
+/// Username and password based authenticator.
 class PasswordAuthenticator implements Authenticator {
   final String username;
   final String password;